Well it's something
It does generate some sort of spec. It only works for catkin at the time
being, since all other packages try to look for catkin itself. I'll have
to make changes to the generator file to get rid of this behaviour. The
generator file is just a copy of the debian generator currently, with
minor tweaks.